+++ R/formulas-sirius.R
@@ -102,12 +102,10 @@ processSIRIUSFormulas <- function(msFName, outPath, adduct, ...)
             
             forms[, rank := NULL]
             
-            # Precursor is always present in MS/MS spectrum: it's added by SIRIUS if necessarily (with zero intensity).
-            # Remove it and use its mz to get ion_formula_mz
-            forms[, ion_formula_mz := mapply(ion_formula, fragInfo,
-                                             FUN = function(form, fi) fi$ion_formula_mz[form == fi$ion_formula])]
             forms[, fragInfo := Map(ion_formula, fragInfo, f = function(form, fi)
             {
+                # Precursor is always present in MS/MS spectrum: it's added by SIRIUS if necessary (with zero intensity).
+                # --> Remove it
                 fi <- fi[intensity != 0 | ion_formula != form]
                 fi[, intensity := NULL]
                 return(fi)